Ensure the latest connectionsDoc from the docSet is used
The error "Cannot pass an old state object to a connection" was due to
the fact that the Server object was maintaining its own copy of the
connections document in the member variable this.connectionsDoc. When a
client sent its copy of the connectionsDoc to the server, this would
update the 'connections' entry in the docSet, but not
this.connectionsDoc. Thus, when subsequently performing
Automerge.change(this.connectionsDoc, ...) the document being changed
was not the latest document in the docSet.

Updated the code to always get the latest document state from the docSet
rather than maintaining a separate member variable.